Separate debug tests into multiple files
