+++ debug/test/browser/debug.test.js
@@ -1,10 +1,9 @@
-import { createElement as h, options, render, createRef, Component, Fragment } from 'preact';
-import { lazy, Suspense } from 'preact/compat';
-import { useState, useEffect, useLayoutEffect, useMemo, useCallback } from 'preact/hooks';
-import { act, setupRerender } from 'preact/test-utils';
-import { setupScratch, teardown, clearOptions, serializeHtml } from '../../../test/_util/helpers';
+import { createElement as h, render, createRef, Component, Fragment } from 'preact';
+import { setupScratch, teardown, serializeHtml } from '../../../test/_util/helpers';
 import { serializeVNode, initDebug } from '../../src/debug';
 import * as PropTypes from 'prop-types';
+
+initDebug();
 
 /** @jsx h */
 
@@ -13,20 +12,12 @@ describe('debug', () => {
 	let errors = [];
 	let warnings = [];
 
-	let diffSpy;
-
 	beforeEach(() => {
 		errors = [];
 		warnings = [];
 		scratch = setupScratch();
 		sinon.stub(console, 'error').callsFake(e => errors.push(e));
 		sinon.stub(console, 'warn').callsFake(w => warnings.push(w));
-
-		clearOptions();
-		diffSpy = sinon.spy();
-		options._diff = diffSpy;
-
-		initDebug();
 	});
 
 	afterEach(() => {
@@ -35,11 +26,6 @@ describe('debug', () => {
 		(console.error).restore();
 		(console.warn).restore();
 		teardown(scratch);
-	});
-
-	it('should call previous options', () => {
-		render(<div />, scratch);
-		expect(diffSpy, 'diff').to.have.been.called;
 	});
 
 	it('should print an error on rendering on undefined parent', () => {
@@ -58,6 +44,7 @@ describe('debug', () => {
 		let fn = () => render(<App />, 6);
 		expect(fn).to.throw(/<App/);
 		expect(fn).to.throw(/6/);
+
 		fn = () => render(<App />, {});
 		expect(fn).to.throw(/<App/);
 		expect(fn).to.throw(/[object Object]/);
@@ -83,6 +70,17 @@ describe('debug', () => {
 		expect(fn).to.throw(/createElement/);
 	});
 
+	it('should print an error when component is an array', () => {
+		let fn = () => render(h([<div />]), scratch);
+		expect(fn).to.throw(/createElement/);
+	});
+
+	it('should print an error on double jsx conversion', () => {
+		let Foo = <div />;
+		let fn = () => render(h(<Foo />), scratch);
+		expect(fn).to.throw(/createElement/);
+	});
+
 	it('should add __source to the vnode in debug mode.', () => {
 		const vnode = h('div', {
 			__source: {
@@ -105,90 +103,6 @@ describe('debug', () => {
 		expect(vnode.props.__self).to.be.undefined;
 	});
 
-	// TODO: Fix this test. It only passed before because App was the first component
-	// into render so currentComponent in hooks/index.js wasn't set yet. However,
-	// any children under App wouldn't have thrown the error if they did what App
-	// did because currentComponent would be set to App.
-	// In other words, hooks never clear currentComponent so once it is set, it won't
-	// be unset
-	it.skip('should throw an error when using a hook outside a render', () => {
-		const Foo = props => props.children;
-		class App extends Component {
-			componentWillMount() {
-				useState();
-			}
-
-			render() {
-				return <p>test</p>;
-			}
-		}
-		const fn = () => act(() => render(<Foo><App /></Foo>, scratch));
-		expect(fn).to.throw(/Hook can only be invoked from render/);
-	});
-
-	// TODO: Fix this test. It only passed before because render was never called.
-	// Once render is called, currentComponent is set and never unset so calls to
-	// hooks outside of components would still work.
-	it.skip('should throw an error when invoked outside of a component', () => {
-		function Foo(props) {
-			useEffect(() => {}); // Pretend to use a hook
-			return props.children;
-		}
-
-		const fn = () => act(() => {
-			render(<Foo>Hello!</Foo>, scratch);
-			useState();
-		});
-		expect(fn).to.throw(/Hook can only be invoked from render/);
-	});
-
-	it('should warn for argumentless useEffect hooks', () => {
-		const App = () => {
-			const [state] = useState('test');
-			useEffect(() => 'test');
-			return (
-				<p>{state}</p>
-			);
-		};
-		render(<App />, scratch);
-		expect(warnings[0]).to.match(/You should provide an array of arguments/);
-		render(<App />, scratch);
-		expect(warnings[1]).to.be.undefined;
-	});
-
-	it('should warn for argumentless useLayoutEffect hooks', () => {
-		const App = () => {
-			const [state] = useState('test');
-			useLayoutEffect(() => 'test');
-			return (
-				<p>{state}</p>
-			);
-		};
-		render(<App />, scratch);
-		expect(warnings[0]).to.match(/You should provide an array of arguments/);
-		render(<App />, scratch);
-		expect(warnings[1]).to.be.undefined;
-	});
-
-	it('should not warn for argumented effect hooks', () => {
-		const App = () => {
-			const [state] = useState('test');
-			useLayoutEffect(() => 'test', []);
-			useEffect(() => 'test', [state]);
-			return (
-				<p>{state}</p>
-			);
-		};
-		const fn = () => act(() => render(<App />, scratch));
-		expect(fn).to.not.throw();
-	});
-
-	it('should print an error on double jsx conversion', () => {
-		let Foo = <div />;
-		let fn = () => render(h(<Foo />), scratch);
-		expect(fn).to.throw(/createElement/);
-	});
-
 	it('should throw errors when accessing certain attributes', () => {
 		const vnode = h('div', null);
 		expect(() => vnode).to.not.throw();
@@ -198,11 +112,6 @@ describe('debug', () => {
 		expect(() => vnode.attributes = {}).to.throw(/use vnode.props/);
 		expect(() => vnode.nodeName = 'test').to.throw(/use vnode.type/);
 		expect(() => vnode.children = [<div />]).to.throw(/use vnode.props.children/);
-	});
-
-	it('should print an error when component is an array', () => {
-		let fn = () => render(h([<div />]), scratch);
-		expect(fn).to.throw(/createElement/);
 	});
 
 	it('should warn when calling setState inside the constructor', () => {
@@ -289,37 +198,9 @@ describe('debug', () => {
 		expect(fn).to.throw(/not valid/);
 	});
 
-	it('should warn for useless useMemo calls', () => {
-		const App = () => {
-			const [people] = useState([40, 20, 60, 80]);
-			const retiredPeople = useMemo(() => people.filter(x => x >= 60));
-			const cb = useCallback(() => () => 'test');
-			return <p onClick={cb}>{retiredPeople.map(x => x)}</p>;
-		};
-		render(<App />, scratch);
-		expect(warnings.length).to.equal(2);
-	});
-
-	it('should warn when non-array args is passed', () => {
-		const App = () => {
-			const foo = useMemo(() => 'foo', 12);
-			return <p>{foo}</p>;
-		};
-		render(<App />, scratch);
-		expect(warnings[0]).to.match(/without passing arguments/);
-	});
-
 	it('should print an error on invalid refs', () => {
 		let fn = () => render(<div ref="a" />, scratch);
 		expect(fn).to.throw(/createRef/);
-
-		// Allow strings for compat
-		let vnode = <div ref="a" />;
-
-		/** @type {*} */
-		(vnode).$$typeof = 'foo';
-		render(vnode, scratch);
-		expect(console.error).to.not.be.called;
 	});
 
 	it('should not print for null as a handler', () => {
@@ -350,20 +231,6 @@ describe('debug', () => {
 		let ref = createRef();
 		render(<div ref={ref} />, scratch);
 		expect(console.error).to.not.be.called;
-	});
-
-	it('should throw an error when missing Suspense', () => {
-		const Foo = () => <div>Foo</div>;
-		const LazyComp = lazy(() => new Promise(resolve => resolve({ default: Foo })));
-		const fn = () => {
-			render((
-				<Fragment>
-					<LazyComp />
-				</Fragment>
-			), scratch);
-		};
-
-		expect(fn).to.throw(/Missing Suspense/gi);
 	});
 
 	describe('duplicate keys', () => {
@@ -634,7 +501,6 @@ describe('debug', () => {
 		});
 	});
 
-
 	describe('PropTypes', () => {
 		it('should fail if props don\'t match prop-types', () => {
 			function Foo(props) {
@@ -678,127 +544,6 @@ describe('debug', () => {
 
 			render(<Bar text="foo" />, scratch);
 			expect(console.error).to.not.be.called;
-		});
-
-		it('should validate propTypes inside lazy()', () => {
-			const rerender = setupRerender();
-
-			function Baz(props) {
-				return <h1>{props.unhappy}</h1>;
-			}
-
-			Baz.propTypes = {
-				unhappy: function alwaysThrows(obj, key) { if (obj[key] === 'signal') throw Error('got prop inside lazy()'); }
-			};
-
-
-			const loader = Promise.resolve({ default: Baz });
-			const LazyBaz = lazy(() => loader);
-
-			const suspense = (
-				<Suspense fallback={<div>fallback...</div>}>
-					<LazyBaz unhappy="signal" />
-				</Suspense>
-			);
-			render(suspense, scratch);
-
-			expect(console.error).to.not.be.called;
-
-			return loader
-				.then(() => Promise.all(suspense._component._suspensions))
-				.then(() => {
-					rerender();
-					expect(errors.length).to.equal(1);
-					expect(errors[0].includes('got prop')).to.equal(true);
-					expect(serializeHtml(scratch)).to.equal('<h1>signal</h1>');
-				});
-		});
-
-		it('should throw on missing <Suspense>', () => {
-			function Foo() {
-				throw Promise.resolve();
-			}
-
-			expect(() => render(<Foo />, scratch)).to.throw;
-		});
-
-		describe('warn for PropTypes on lazy()', () => {
-			it('should log the function name', () => {
-				const loader = Promise.resolve({ default: function MyLazyLoadedComponent() { return <div>Hi there</div>; } });
-				const FakeLazy = lazy(() => loader);
-				FakeLazy.propTypes = {};
-				const suspense = (
-					<Suspense fallback={<div>fallback...</div>} >
-						<FakeLazy />
-					</Suspense>
-				);
-				render(suspense, scratch);
-
-				return loader
-					.then(() => Promise.all(suspense._component._suspensions))
-					.then(() => {
-						expect(console.warn).to.be.calledTwice;
-						expect(warnings[1].includes('MyLazyLoadedComponent')).to.equal(true);
-					});
-			});
-
-			it('should log the displayName', () => {
-				function MyLazyLoadedComponent() { return <div>Hi there</div>; }
-				MyLazyLoadedComponent.displayName = 'HelloLazy';
-				const loader = Promise.resolve({ default: MyLazyLoadedComponent });
-				const FakeLazy = lazy(() => loader);
-				FakeLazy.propTypes = {};
-				const suspense = (
-					<Suspense fallback={<div>fallback...</div>} >
-						<FakeLazy />
-					</Suspense>
-				);
-				render(suspense, scratch);
-
-				return loader
-					.then(() => Promise.all(suspense._component._suspensions))
-					.then(() => {
-						expect(console.warn).to.be.calledTwice;
-						expect(warnings[1].includes('HelloLazy')).to.equal(true);
-					});
-			});
-
-			it('should not log a component if lazy throws', () => {
-				const loader = Promise.reject(new Error('Hey there'));
-				const FakeLazy = lazy(() => loader);
-				FakeLazy.propTypes = {};
-				render(
-					<Suspense fallback={<div>fallback...</div>} >
-						<FakeLazy />
-					</Suspense>,
-					scratch
-				);
-
-				return loader.catch(() => {
-					expect(console.warn).to.be.calledOnce;
-				});
-			});
-
-			it('should not log a component if lazy\'s loader throws', () => {
-				const FakeLazy = lazy(() => { throw new Error('Hello'); });
-				FakeLazy.propTypes = {};
-				let error;
-				try {
-					render(
-						<Suspense fallback={<div>fallback...</div>} >
-							<FakeLazy />
-						</Suspense>,
-						scratch
-					);
-				}
-				catch (e) {
-					error = e;
-				}
-
-				expect(console.warn).to.be.calledOnce;
-				expect(error).not.to.be.undefined;
-				expect(error.message).to.eql('Hello');
-			});
 		});
 	});
 });
